refactor(webview): move properties to WebViewWindowMixin (Iteration #129)
